Moved App.tsx to src directory
