+++ mobile/src/components/Button/index.js
@@ -1,0 +1,3 @@
+import styled from 'styled-components/native';
+
+export const Button = styled.button``;
